Update module imports Move config file
