+++ src/config/config.ts
@@ -20,3 +20,5 @@ export const CACHE_SERVER = getEnvironmentValue("CACHE_SERVER");
 export const API_URL = getEnvironmentValue("API_URL");
 
 export const CHS_URL = getEnvironmentValue("CHS_URL");
+
+export const APPLICATION_NAME = "certificates.orders.web.ch.gov.uk";
