+++ src/test/controllers/certificate.options.spec.unit.ts
@@ -1,4 +1,4 @@
-import {setItemOptions} from "../../controllers/certificate.options.controller";
+import { setItemOptions } from "../../controllers/certificate.options.controller";
 
 describe("setItemOptions", () => {
